pyproject.toml
[project]
name = "nestlet"
version = "0.1.0"
description = "A lightweight neuro-symbolic AI agent with recursive reasoning and autonomous tool-making"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Nestlet Contributors" }
]
keywords = ["agent", "llm", "neuro-symbolic", "ai", "tool-making"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "openai>=1.42.0",           # OpenAI API
    "anthropic>=0.34.0",        # Anthropic API
    "pydantic>=2.8.0",          # Data validation
    "pyyaml>=6.0",              # YAML configuration
    "ipykernel>=6.26.0",        # Python Kernel
    "pandas>=2.0.0",            # Data processing
    "numpy>=1.24.0",            # Numerical computation
    "aiofiles>=24.1.0",         # Async file operations
    "aiohttp>=3.9.0",           # Async HTTP client (for web search)
    "rich>=13.0.0",             # Terminal beautification
    # Office document parsing
    "python-docx>=1.1.0",       # Word documents
    "openpyxl>=3.1.0",          # Excel spreadsheets (.xlsx)
    "xlrd>=2.0.1",              # Excel spreadsheets (.xls, legacy OLE2 format)
    "olefile>=0.47",            # OLE2 file analysis (for OneDrive placeholder detection)
    "python-pptx>=0.6.23",      # PowerPoint presentations
    "pypdf>=4.0.0",             # PDF documents
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
azure = [
    "azure-identity>=1.15.0",
]
search = [
    "duckduckgo-search>=4.0.0",  # DuckDuckGo search backend (optional fallback)
]

[project.scripts]
nestlet = "inception.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["inception"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
